requirements.txt
requests
bs4
lxml
img2pdf
humanfriendly
motor
python-dateutil